refactor(simd): split utf8.rs into a module directory
