+++ docs/adrs/adr-0013.md
@@ -17,7 +17,7 @@ UTF-8-heavy strings and **0.86–0.91×** (a regression) on ASCII-heavy structur
 Three of the issue's stated premises had gone stale by the time it was picked up:
 
 - The AVX2 Keiser–Lemire UTF-8 validator it planned to salvage from PR #127 had already landed,
-  better and differentially tested, in [`src/text/utf8.rs`](../../src/text/utf8.rs).
+  better and differentially tested, in [`src/text/utf8/mod.rs`](../../src/text/utf8/mod.rs).
 - The quote/backslash/control scanners it planned to salvage already existed as
   [`find_json_escape`](../../src/util/simd/escape.rs), across NEON/AVX2/SSE2/scalar.
 - The `NestingStack` was justified as fixing the #151 recursion DoS, but #151 was already fixed by
